CLI now accepts and displays punctuation characters and symbols #5
TI spent wayyyyy more time on this tahn it seems i needed to. I was trying to use "String.fromCharCode(keyCode)" but after a lot of searching I found that special characters don't have direct ASCII mappings like letters and numbers do. So this is what I came up with