Implementing command completion #5
I thought about using an array to store all the shell commands and then depending on if the first few characters typed match up to one of the shell commands in the array it would complete. However was having a hard time implementing this. I tried looking up a few solutions, with this one being the most strait forward and "clean". I understand what it is doing and how. I wouldn't have been able to write it without help due to my limited knowledge of Typescript capabilities.