Implemented line scrolling #4
I was working on this for a little bit (a much more complicated way), but found a site (https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/putImageData) that had a description of a ton of built in function descriptions and found there was a much easier way to do that. Your discussion in class really helped. That being just grabbing the image and just repasting it once you exceed the length of the canvas.

I also took "<string>" out of two of the shell commands. I originally added them because I was just looking at the ones above and saw they had it, but I've since realized why.