Develop a memory accessor prototype and implement it in the host OS #18
The majority of this code has to do with the development of the memory accessor

I have two files. one for memory and the other for memory accessor. memory is for the actual memory that the computer will be manipulating. memory accessor is the accessor. self explanatory, but like the classic meme, you know its a birch tree because the way that it is.

I also just created a file and barely built it out for the memory manager. but thats to be worked on later