Built out the PCB #15, developed the memory accessor #18 and actively buidling the manager #16 and cpu prototype #19
I thought I had pushed my work towards the accessor, but I guess not. Sorry for the rather large commit. I try to stay away from this, but at times I end up having a larger commit than usual. I'll continue to work on this.

I developed the key components for code to run on the os. I am still actively working on the CPU prototype and the memory manager.

I also hope my branch is working properly. I've never used branches before and I watched a video on it an I see that I am commiting, but not shure if i have to create a pull request on github for them to go through.