Implemented the ability to load a program into memory #10
Wow. you would not believe the headach i had getting this to work and the laughter i had when i remembered that I had to add the scripts into the index.html file. I almost killed my roomate (not really, but you get the point)

I made some changes to the kernel and mainly the shell file to allow the user program inputs to be stored in memory. My comments should give a good description on how i was able to accomplish this