A few implementations. See description to read what was added #19 #11 #12
I apologize for a rather large commit. A lot of it had to do with running a program where I ended up having to fiddle with a few classes. A description on what this commit includes is numbered below.

1) the contribution to num19 (develop a cpu prototype) includes changes to the cycle() method and my instruction set.

The addition/change to my cycle() method was debugging. I got rid of that error catching because it just doesn't make sense. this way is also easier to read in the console.

I ended up changing my instructions to match the ones from the link you had provided. I also made them match how I had them written in org and arch, dealing with a high and low order bytes. I'm not sure if it's needed, but 1 i was actually able to run a program, and two I was experiencing less errors. Some programs don't fully run but my registers update and hold values, and it matches similarly when I test on hall of fame projects. I'm still working through that but I figured I should push before I start adding more.

2) I added the command run<pid> num11 in issues to run a program in memory. I haven't gotten this quite yet, BUT I am able to execute several lines. A better description on the issue was above. The run command takes a pid and runs that specified program. I also added a mehtod to desiplay the memory. You can change the value that i is < than to show more. This is just so I could see before I added the visual to the ui.

3) For num12 (execute a running program) I can execute a running program, but not fully. Like i previously stated, I can execute several lines and update registers, but I am still working on fine tuning instructions (which I think is where my error lies). I have debugging to show the states of the register but they are show in the console itself. I am not planning on keeping them there, but I want to commit this before adding the display of those on the ui.