Implemented line wrap #14
I've spent a good amount of time on this trying to perfect it. Line wrapping works great. The only issue I still have is deleting characters. I can delete them like normal, but when going from the new line to the previous the x value seems to get screwed up so sometimes half a character will be deleted instead of the entire thing but it's never substantial. I mean how often will we be using backspace for text going past the console limit?