Users can load three programs into memory at once along with storing each with their own partition
Some of my .js files that are updated may because of the fact that I merged iproject2 into the main branch.

#22:
This is included because the work for storing multiple programs in memory had to be done in order for the user to load three programs into memory at once. I may have to come back to this after confirming with a classmate that you mean just typing load three times rather than typing load in one string and then pressing enter, but I should be good.

#25:
I enabled the ability for a user to store multiple programs in memory. Each program has it's own segment allocated by the client OS which keeps track of partitions used.

My memory display has updated as expected because we increased the memory size (768) along with confirming the programs are being saved in their respective spots.