Implemented PCB display #24 Added the out of bounds debug #27
Major addition is the PCB is now displayed. It isn't fully functional as I havn't added the ability to assign a process a que along with scheduling so as of now it's to confirm a) that i can display the pcb and b) that the program is being stored in the correct partitions

I also just aded a kernel output to let you know if the program counter has or will exceed memory bounds. I also had to change an output. I was using throw new error instead of krnTrace. It was probably becaues I had to use throw new error to stop a program from going into memory because I couldn't figure it out. That had been resolved in the last commit.