Note! (base and limit are slightly off)
I forgot to mention in my last two commits that the base and limits were a place or two wrong. So there most definitely  will be an update to that in the near future. It was fine before so if all fails I'l just go back to a previous file.