Working to develop the scheduler #30 and implemented the "runall" command #23
I've been working on creating a round robin scheduler with a q of  6.

I'm able to context switch between programs as they execute with the "runall" command I Implemented.

I have the cpu communicating directly with the memory manager (which I know I shouldn't do, but for the purposes of testing I have it set up to iterate through the readyQueue).

One note is that my instruction set is still a little wacky so the program counter may seem greater than what it should be, but the context switch is only being called after 6 cpu cycles.

Last note is that I am aware that this is supposed to be in the client os. I unfortionitely missed that part and put it in my cpu class so I'll have to move some things around. But I figured I mind as well commit this so it doesn't look like i've been doing nothing.