I updated the kill<pid> command #23
I had to alter this along with adding some checks within the scheduler to ensure that if you ran kill<pid> all other processes would continue to run. Before it would stop the entire cycle method.

I also updated how we handle the quantum for scheduling. I found that there was an error when trying to change it's value. It just so happened that in testing, I chose the number that it was stuck too. #30