Fixing memory bounds #26
I found that if I were to upload a program that exceeded the partition size it would output in the terminal that it exceeded, but it would still load it in.

I have since made it so if a program exceeds the partition size, it isn't loaded in and an error message is output.