Updated my scheduler and how i handle context switching #30
I was reading/remembered that we needed to handle context switching and dispatching in the kernel. I've now added a dispatcher and scheduler into the kernel which can communicate with the scheduler.ts file. #31

I also changed how I was stopping a program/knew when one was complete. Before since I couldn't think of something I just had the program terminate once the next instruction was 0x00, it now just checks if there is another instruction to execute at all using if(!instruction). Not sure how I didn't think of that before.

Lastly I wanted to note that while my instruction set still doesn't work fully (I think it's an issue with timing and partially my brach) my scheduler still works and can context switch when needed. I can also output "inner1" when running the larger command on your website, but am aware that it doesn't fully run through. I'd like to get this working for the final project, but I wanted to spend the most time on this and the memory partitioning as that was the main part of the project.