Really tried to fix enforcing memory bounds #27
I've been trying to trouble shoot this all day. I have what I believe should be the correct way to enforce memory bounds commented out. This is because while it enforces bounds when running ONE program, it has trouble when context switching is brought into the mix. Once a context switch takes place, an instruction or two will execute, and then I'll get a memory out of bounds at address 236. It will say this for both process 1 and 2. I had a lot of debugging and found that when a context switch occurs the base and limit of the pids is updating per the process in execution, but an out of bounds error at 236 is still picked up somewhere. I've just run out of ideas for what it could be