Implemented a disk system viewer #43
I was able to not only add the disk system display, but in turn it let me verify that my initialization of the disk system worked.

I also updated the disk driver because I thought i had to have 4 tracks and 8 for the sector and block, but in reality i needed to go one down for each. #44

I now have to make it so format can be called using the shell to initialize the disk rather than initializing it on load.