Added the format shell command #35
I added the format command within the shell to format the disk driver. I saw some hall of fame projects that required you to format the disk driver for it to be initialized so that's what I did. Honestly I may revert to how I handled initialization before, but for now I'll keep it as is