updated the write function #38
I found that I could still write to a file if the page was re-freshed and i hadn't formatted the disk. I ensured that it uses the flag that I set up earlier to ensure the disk has been formatted

Also the way gitHub presents this change seems weird. not a fan