Updated the write function #44
After a lot of trial and error I fixed the write method. Now when we write to a file name after data has already been written to it, the previous data is overwritten.

I am aware that on a modern disk the data isn't cleared if we don't write to that block, it just becomes available, but it was easier for me to do it this way and I know you mentioned in class that it is ok to do it either way. Especially since some hall of fame projects had it like this

I should probably put a check in the create function to ensure that i don't have two of the same filenames. I thought of that as I was writing this