Partially implemented the ability to roll out a process
This is the first steps I've made to tackle the ability to swap virtual memory with enough physical memory for three concurrent suer processes.

I slightly updated my pcb to track whether a process in on disk or not.

I added a roll out function which I havn't fully tested yet or set up the ability to, but I think in theory it should work. I just don't want to end up committing a giant file. So this is a piece of the puzzle