Added the ability to roll in a process #44
I added a roll out function in the dsDD and still need to check it but like i said with the roll out i think in theory it should work

The other thing I implemented was the ability to load multiple programs to memory. If there are more than three loaded into memory, the others go onto the disk. it should be noted at this point that it doesn't exactly work in the sense that the information loaded isn't exactly correct, but I believe I know why. Im glad to see that I am able to load a fourth process onto the disk though.