Added more info for user to see what process was loaded #45
I added more information about where the file was being saved and to what name when the number of partitions ran out.

I also tried to make it so when reading the code was displayed through several lines. As of right now no matter how long the text string may be it is going over the canvas size. I tried looking at why that is and how to fix it but still nothing. I figured I'd leave it as is for now as it isn't too important. At least we can see the first 20 some characters to confirm the loaded program is correct.

I have to change how the code is being displayed in the file system to reflect its true data