Additions to loading a program onto disk #45
I had been working on trying to allow the user to read the loaded user program contents but it was too troublesome. Either I was able to read the user program input and not file contents that were added by "write aidan "hello world" as that would just be turned into hex and vise versa. I added a check at the start of the write function that says if the filename has the word process in it (which all will) it is unable to be read. I hope this is ok because it doesn't specifically say anything about needing this in the project outline and the majority of hall of fame projects didn't let you read loaded user programs either.

I also updated the memory manager load method to add the loaded process to the resident list and update it's information