Wrote roll-in and roll-out routines #45
I implemented roll in and roll out routines in the memory manager which uses the disk driver to store a process to the disk. I also added a function to swap the desired process

This isn't 110% working. It works, but there are a few heisenbugs.

I am able to run the fith process by taking the first process and swapping it with the desired (fifth) to run. I can then run the process that was swapped out of memory. The issue arrises when trying to swap more than one concurrently as in put two files that were previously in memory onto the disk and then trying to swap them back into memory. I think i have a solution but we'll have to wait and see