Removed the ability to delete a process #39
So im an idiot.

I found that because i made it so deleting a file meant that it was removed from the pid, in my swapper i was using delete file to remove it's contents. so i just totally got rid of my previous update to the delete method. The swapper works again