Worked on allowing a user to run x number of processes #45
I updated the scheduler a bit in order for processes that are on the disk to be able ot be swapped out and scheduled accordingly.

This isn't perfect. At the end when there is only one or two processes left it spazzes and goes back and forth and doesn't allow one to complete.

I was fiddling around with the run all command for a bit but that wasn't working (I know that when running all no matter if there are processes on disk or not it runs the first process twice. I have to look into this more.

NOTE: this only works for 4 processes loaded, not 5.  theres a weird issue that i was working on trying to fix last night where if there were more than one processes on the disk it was having trouble swapping. I'm not sure what the root of this issue is but it's just so darn annoying. That of course is what is causing the issues when there is more than one process on a disk.

It should be said that I can run thorugh all processes one a time including when a swap is needed.